requirements.txt
aiopg==1.4.0
aiosqlite==0.20.0
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.0.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
idna==3.10
iniconfig==2.0.0
jose==1.0.0
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
python-jose==3.3.0
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
starlette==0.38.6
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
